Add: 맛집 상세 지도 API 호출 / 평점

+++ src/main/java/com/wisein/wiselab/dao/MatzipDAO.java
@@ -24,6 +24,6 @@ public interface MatzipDAO {
 
     public int recmCnt(int matzipId) throws Exception;
 
-    public int avgRate(int matzipId) throws Exception;
+    public float avgRate(int matzipId) throws Exception;
 
 }
